f2fs: use new shrinker API

+++ fs/f2fs/f2fs.h
@@ -2000,8 +2000,8 @@ int f2fs_read_inline_dir(struct file *, void *, filldir_t, struct f2fs_str *);
 /*
  * shrinker.c
  */
-int f2fs_shrink_count(struct shrinker *, struct shrink_control *);
-int f2fs_shrink_scan(struct shrinker *, struct shrink_control *);
+unsigned long f2fs_shrink_count(struct shrinker *, struct shrink_control *);
+unsigned long f2fs_shrink_scan(struct shrinker *, struct shrink_control *);
 void f2fs_join_shrinker(struct f2fs_sb_info *);
 void f2fs_leave_shrinker(struct f2fs_sb_info *);
